+++ artemis/src/main/java/com/artemis/InvocationStrategy.java
@@ -11,13 +11,12 @@ public class InvocationStrategy extends SystemInvocationStrategy {
 	/** Processes all systems in order. */
 	@Override
 	protected void process(Bag<BaseSystem> systems) {
-		Object[] systemsData = systems.getData();
+		BaseSystem[] systemsData = systems.getData();
 		for (int i = 0, s = systems.size(); s > i; i++) {
 			if (disabled.get(i))
 				continue;
 
-			BaseSystem system = (BaseSystem) systemsData[i];
-			system.process();
+			systemsData[i].process();
 			updateEntityStates();
 		}
 	}